requirements.txt
pdfplumber
langchain
openai
faiss-cpu
streamlit
numpy
tiktoken
requests
beautifulsoup4